Fix #760
.recover_data_from_environment() was meant to get the environment from
the model formula, if available, but failed to do so. Extract the right
formula from find_formula() so the environment can be used correctly.

Additionally, if there is an object of the same name in the parent
environment, eval() would look there, and we'd use that before
considering the model environment. We want to consider the parent frame
last, so force eval() not to do this.