Do not fall back to parent.frame()
The parent.frame() of. .recover_data_from_environment() is always
another function inside insight, since this is an internal non-exported
function. The user's data shouldn't be accessible from here, and in the
rare case when the user's data has the same name as a local variable in
an insight function, this could cause very confusing results.