Fix get_predicted.lm()'s handling of scaled columns
Scaled columns (e.g. with scale() in a model formula) become matrices,
not numeric, causing problems when passing data to predict() that is
numeric. In #671, get_predicted.lm() was changed to detect this and
convert the affected columns to numeric.

This worked by modifying the model frame directly, fixing the types of
the affected columns. But then .get_predicted_args() was used to get the
arguments to use, including the data; and if get_data() successfully
located the data, it would not use the model frame. This hence only
worked if get_data() fell back to the model frame. If the data was in
the global environment, for instance, get_data() would find it there,
not fall back to model.frame(), and cause predicting to fail.

With the fixes to get_data() in #761, get_data() locates the data
successfully more often, making the fix fail. Adjust get_predicted.lm()
to fix the actual data that will be passed to predict(), not just the
model frame, so this problem is avoided. Adjust the test to consider
more circumstances.